+++ src/connector.c
@@ -85,7 +85,7 @@ static int ssh_connector_channel_data_cb(ssh_session session,
                                          void *userdata);
 static int ssh_connector_channel_write_wontblock_cb(ssh_session session,
                                                     ssh_channel channel,
-                                                    size_t bytes,
+                                                    uint32_t bytes,
                                                     void *userdata);
 static ssize_t ssh_connector_fd_read(ssh_connector connector,
                                      void *buffer,
@@ -248,14 +248,14 @@ static void ssh_connector_fd_in_cb(ssh_connector connector)
     uint32_t toread = CHUNKSIZE;
     ssize_t r;
     ssize_t w;
-    int total = 0;
+    ssize_t total = 0;
     int rc;
 
     SSH_LOG(SSH_LOG_TRACE, "connector POLLIN event for fd %d", connector->in_fd);
 
     if (connector->out_wontblock) {
         if (connector->out_channel != NULL) {
-            size_t size = ssh_channel_window_size(connector->out_channel);
+            uint32_t size = ssh_channel_window_size(connector->out_channel);
 
             /* Don't attempt reading more than the window */
             toread = MIN(size, CHUNKSIZE);
@@ -328,9 +328,9 @@ static void ssh_connector_fd_in_cb(ssh_connector connector)
  */
 static void ssh_connector_fd_out_cb(ssh_connector connector){
     unsigned char buffer[CHUNKSIZE];
-    int r;
-    int w;
-    int total = 0;
+    ssize_t r;
+    ssize_t w;
+    ssize_t total = 0;
     SSH_LOG(SSH_LOG_TRACE, "connector POLLOUT event for fd %d", connector->out_fd);
 
     if(connector->in_available){
@@ -432,7 +432,7 @@ static int ssh_connector_channel_data_cb(ssh_session session,
 {
     ssh_connector connector = userdata;
     int w;
-    size_t window;
+    uint32_t window;
 
     (void) session;
     (void) channel;
@@ -453,7 +453,7 @@ static int ssh_connector_channel_data_cb(ssh_session session,
 
     if (connector->out_wontblock) {
         if (connector->out_channel != NULL) {
-            int window_len;
+            uint32_t window_len;
 
             window = ssh_channel_window_size(connector->out_channel);
             window_len = MIN(window, len);
@@ -517,7 +517,7 @@ static int ssh_connector_channel_data_cb(ssh_session session,
  */
 static int ssh_connector_channel_write_wontblock_cb(ssh_session session,
                                                     ssh_channel channel,
-                                                    size_t bytes,
+                                                    uint32_t bytes,
                                                     void *userdata)
 {
     ssh_connector connector = userdata;
@@ -529,7 +529,7 @@ static int ssh_connector_channel_write_wontblock_cb(ssh_session session,
     SSH_LOG(SSH_LOG_TRACE, "Channel write won't block");
     if (connector->in_available) {
         if (connector->in_channel != NULL) {
-            size_t len = MIN(CHUNKSIZE, bytes);
+            uint32_t len = MIN(CHUNKSIZE, bytes);
 
             r = ssh_channel_read_nonblocking(connector->in_channel,
                                              buffer,
